duplicate_BIDS_analysis refactored and uses logs
 Changes to be committed:
	deleted:    duplicate_BIDS_analysis.py
	new file:   duplicate_BIDS_analysis/File_locate_info.py
	new file:   duplicate_BIDS_analysis/__init__.py
	new file:   duplicate_BIDS_analysis/find.py
	new file:   duplicate_BIDS_analysis/main.py
	new file:   duplicate_BIDS_analysis/rename.py

Call the function with:

  from duplicate_BIDS_analysis.main import run as run_dba

  run_dba(api_key, project_path, do_make_changes, do_use_latest_not_largest, verbose)
